# routes/api.yaml
swagger: "2.0"
info:
  description: "This will serve as the official API documentation for Cockpit."
  version: "2.0.0"
  title: "COCKPIT"
  termsOfService: ""
  contact:
    email: "juan.navarrete@meggitt.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost5000"
basePath: "/v1"
...
schemes:
  - "https"
  - "http"
paths:
...
  /api/profile:
    post:
      tags:
        - "/api/profile"
      summary:  "Create or update user profile"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      description: "To create or update profile"
      operationId: " POST /api/profile"
      parameters:
        - name: "id"
          in: "body"
          description: "profile identifier"
          required: false
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Profile"
      responses:
        200:
          description: "ok"
        400:
          description: "Error"
        500:
          description: "Internal server error"
    get:
      tags:
        - "/api/profile"
      summary: "Get all profiles"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      description: "To retrieve all profiles"
      operationId: "GET /api/profile"
      parameters:
        - name: "id"
          in: "body"
          description: "To retrieve all profiles"
          required: true
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Profile"
      responses:
        200:
          description: "ok"
        400:
          description: "Error"
        500:
          description: "Internal server error"
    delete:
      tags:
        - "/api/profile"
      summary: "Delete profile, user & post"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      description: "To delete a profile"
      operationId: "DELETE /api/profile"
      parameters:
        - name: "id"
          in: "body"
          description: "To retrieve all profiles"
          required: true
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Profile"
      responses:
        200:
          description: "ok"
        400:
          description: "Error"
        500:
          description: "Internal server error"
  /api/profile/experience:
    put:
      tags:
        - "/api/profile/experience"
      summary: "Add profile experience"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      description: "Add experience to profile"
      operationId: "PUT /api/profile/experience"
      parameters:
        - name: "id"
          in: "body"
          description: "Add experience to profile"
          required: true
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Profile"
      responses:
        200:
          description: "ok"
        400:
          description: "Error"
        500:
          description: "Internal server error"
...
  /api/profile/me:
    get:
      tags:
        - "api/profile/me"
      summary: "Get current user's profile"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      description: "get user's profile"
      operationId: "get me"
      parameters:
        - name: "id"
          in: "body"
          description: "Profile identifier"
          required: true
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Profile"
      responses:
        200:
          description: "ok"
        500:
          description: "Internal server error"
...
  /api/post:
    post:
      tags:
        - "post"
      summary: "Create Post"
      description: "Post object."
      operationId: "create Post"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Created Post object"
          required: true
          schema:
            $ref: "#/definitions/Post"
      responses:
        201:
          description: "New user Post"
        400:
          description: "Error"
    get:
      tags:
        - "post"
      summary: "get Post"
      description: "Get Post"
      operationId: "get Post"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Get posts"
          required: false
          schema:
            $ref: "#/definitions/Post"
      responses:
        200:
          description: "ok"
        500:
          description: "Error"
...
definitions:
...
  Profile:
    type: "object"
    properties:
      id:
        type: string
        format: moongoose.object.id()
      user:
        type: object
        format: User
      text:
        type: string
      name:
        type: string
      avatar:
        type: string
      likes:
        type: string
      comments:
        type: string
      date:
        type: string
        format: YYYY-MM-DD HH:MM:SS:mmmZ
  Post:
    type: "object"
    properties:
      id:
        type: string
        format: moongose.object.id()
      user:
        type: object
        format: User
      text:
        type: string
      name:
        type: string
      avatar:
        type: string
      like:
        type: string
      comments:
        type: string
      date:
        type: string
        format: YYYY-MM-DD HH:MM:SS:mmmZ
...