# routes/api/posts.js
const express = require('express');
const router = express.Router();
const { check, validationResult } = require('express-validator/check');
const auth = require('../../middleware/auth');

const Post  = require('../../models/Post');
const Profile = require('../../models/Profile');
const User = require('../../models/User');

//@route    POST api/posts
//@desc     Create a post
//@access   Private
router.post('/',
    [   auth,
        check('text', 'Text is required').not().isEmpty()
    ],
    async (req, res)=>{
        //Check for errors
        const errors = validationResult(req);
        if(!errors.isEmpty()){
            return res.status(400).json({ errors: errors.array() });
        }
        try{
            const user = await User.findById(req.user.id).select('-password');

            const newPost = new Post({
                text: req.body.text,
                name: user.name,
                avatar: user.avatar,
                user: req.user.id
            });

            const post = await newPost.save();

            res.json(post);
        }catch (e) {
            console.error(e.message);
            res.status(500).send('Server Error');
        }
});

//@route    GET api/posts
//@desc     Get all posts
//@access   Private
router.get('/', auth, async (req, res)=>{
   try{
        // Get post and sort
       const posts = await Post.find().sort({ date: -1});  //Get most descending order
       res.json(posts);
   } catch(e){
       console.error(err.message);
       res.status(500).send('Server Error');
   }
});

//@route    GET api/posts/:id
//@desc     Get post by id
//@access   Private
router.get('/:id', auth, async (req, res)=>{
    try{
        // Get post and sort
        const post = await Post.findById(req.params.id);
        if(!post){
            return res.status(404).json({msg: 'Post not found'})
        }
        res.json(post);
    } catch(e){
        console.error(err.message);
        if(e.kind === 'ObjectId'){
            return res.status(404).json({msg: 'Post not found'})
        }
        res.status(500).send('Server Error');
    }
});

//@route    DELETE api/posts/:id
//@desc     delete post by id
//@access   Private
router.delete('/:id', auth, async (req, res)=>{
    try{
        // Get post and sort
        const post = await Post.findById(req.params.id);

        if(!post){
            return res.status(404).json({msg: 'Post not found'})
        }

        //Check the user
        if(post.user.toString() !== req.user.id){
            return res.status(401).json({ msg: 'User is not Authorize'})
        }

        await post.remove();

        res.json({ msg: 'Post removed' });
    } catch(e){
        console.error(err.message);
        if(e.kind === 'ObjectId'){
            return res.status(404).json({msg: 'Post not found'})
        }
        res.status(500).send('Server Error');
    }
});

//@route    Put api/posts/like/:id
//@desc     Like a post
//@access   Private
router.put('/like/:id', auth, async (req, res)=>{
    try{
        const post = await Post.findById(req.params.id);
        if(!post){
            return res.status(404).json({msg: 'Post not found'})
        }

        // Check if the post has already been liked
        if(post.likes.filter(like => like.user.toString() === req.user.id).length > 0 ){
            return res.status(400).json({ msg: 'Post already like'});
        }

        post.likes.unshift({user: req.user.id });

        await post.save();
        res.json(post.likes);
    }catch (e) {
        console.error(e.message);
        res.status(500).send('Server Error');
    }
});

//@route    Put api/posts/unlike/:id
//@desc     Like a post
//@access   Private
router.put('/unlike/:id', auth, async (req, res)=>{
    try{
        const post = await Post.findById(req.params.id);
        if(!post){
            return res.status(404).json({msg: 'Post not found'})
        }

        // Check if the post has already been liked
        if(post.likes.filter(like => like.user.toString() === req.user.id).length === 0 ){
            return res.status(400).json({ msg: 'Post has not yet been liked'});
        }

        //Get remove index
        const removeIndex = post.likes.map(like => like.user.toString()).indexOf(req.user.id);

        post.likes.splice(removeIndex, 1);

        await post.save();

        res.json(post.likes);
    }catch (e) {
        console.error(e.message);
        res.status(500).send('Server Error');
    }
});

//@route    POST api/posts/comment/:id
//@desc     Comment on a post
//@access   Private
router.post('/comment/:id',
    [   auth,
        check('text', 'Text is required').not().isEmpty()
    ],
    async (req, res)=>{
        //Check for errors
        const errors = validationResult(req);
        if(!errors.isEmpty()){
            return res.status(400).json({ errors: errors.array() });
        }
        try{
            const user = await User.findById(req.user.id).select('-password');

            const post = await Post.findById(req.params.id);
            if(!post){
                return res.status(404).json({msg: 'Post not found'})
            }

            const newComment = new Post({
                text: req.body.text,
                name: user.name,
                avatar: user.avatar,
                user: req.user.id
            });

            post.comments.unshift(newComment);

            await post.save();

            res.json(post.comments);
        }catch (e) {
            console.error(e.message);
            res.status(500).send('Server Error');
        }
    });

//@route    DELETE api/posts/comment/:id/:comment_id
//@desc     Delete Comment
//@access   Private
router.delete('/comment/:id/:comment_id', auth, async (req, res) =>{
    try{
        const post = await Post.findById(req.params.id);
        if(!post){
            return res.status(404).json({msg: 'Post not found'})
        }

        // Pull out comment
        const comment = post.comments.find(comment => comment.id === req.params.comment_id);

        // Mare sure comment exits
        if(!comment){
            return res.status(404).json({msg: 'Comment not found'})
        }

        //Check user
        if(comment.user.toString() !== req.user.id){
            return res.status(401).json({msg: 'User not authorized'})
        }

        //Get remove index
        const removeIndex = post.comments.map(comment => comment.user.toString()).indexOf(req.user.id);

        post.comments.splice(removeIndex, 1);

        await post.save();

        res.json(post.comments);
    }catch (e) {
        console.error(e.message);
        res.status(500).send('Sever Error');
    }
});

module.exports = router;