# routes/api/profile.js
const express = require('express');
const request = require('request');
const config = require('config');
const router = express.Router();
const auth = require('../../middleware/auth');
const { check, validationResult } = require('express-validator/check');
const Profile = require('../../models/Profile');
const User = require('../../models/User');
const Post = require('../../models/Post');

//@route    GET api/profile/me
//@desc     Get current user's profile
//@access   Private
router.get('/me', auth, async (req, res)=> {
    try{
        // GET current profile based on current user's id.
        const profile  = await Profile.findOne({ user: req.user.id }).populate('User', ['name', 'avatar']);

        if(!profile){
            return res.status(400).json({ msg: 'There is no profile for this user'})
        }

        res.status(200).json(profile);
    }catch (e) {
        console.error(err.message);
        res.status(500).send('Server Error');
    }
});

//@route    POST api/profile
//@desc     Create or update user profile
//@access   Private
router.post('/', [
    auth,
    [
        check('status', 'Status is required').not().isEmpty(),
        check('skills', 'Skills is required').not().isEmpty()
    ]
],async (req, res) => {
    // checks if there are errors with the Profile model, status and skills are required fields
    const errors = validationResult(req);
    if(!errors.isEmpty()) {
        return res.status(400).json({errors: errors.array()})
    }

    // Build profile object
    const { company, website, location, bio, status, githubusername, skills, youtube, facebook, twitter,
        instagram, linkedin } = req.body;

    const profileFields = {};
    profileFields.user = req.user.id;

    if(company) profileFields.company = company;
    if(website) profileFields.website = website;
    if(location) profileFields.location = location;
    if(bio) profileFields.bio = bio;
    if(status) profileFields.status = status;
    if(githubusername) profileFields.githubusername = githubusername;
    if(skills){
        profileFields.skills = skills.split(',').map(skill=> skill.trim());
    }

    // Build social object
    profileFields.social = {};

    if(youtube) profileFields.social.youtube = youtube;
    if(twitter) profileFields.social.twitter = twitter;
    if(facebook) profileFields.social.facebook = facebook;
    if(linkedin) profileFields.social.linkedin = linkedin;
    if(instagram) profileFields.social.instagram = instagram;

    try{
        let profile = await Profile.findOne({ user: req.user.id });

        if(profile){
            //Update Profile
            profile = await Profile.findOneAndUpdate(
                { user : req.user.id},
                { $set : profileFields},
                { new: true}
                );

            return res.json(profile);
        }

        // Create a Profile
        profile = new Profile(profileFields);

        await profile.save();
        res.status(201).json(profile);
    }catch(e){
        console.error(e.message);
        res.status(500).send('Server Error')
    }
});

//@route    GET api/profile
//@desc     Get all profiles
//@access   Public
router.get('/', async (req, res)=>{
   try {
       const profiles = await Profile.find().populate('user', ['name', 'avatar']);
       res.json(profiles);
   }catch (e) {
       console.error(e.message);
       res.status(500).send('Server Error');
   }
});

//@route    GET api/profile/user/:user_id
//@desc     Get profile by user ID
//@access   Public
router.get('/user/:user_id', async (req, res)=>{
    try {
        const user = req.params.user_id;
        const profile = await Profile.findOne({ user }).populate('user', ['name', 'avatar']);

        if(!profile){
            return res.status(404).json({ msg: 'Profile not found'});
        }
        res.json(profile);
    }catch (e) {
        console.error(e.message);
        if(e.kind == 'ObjectId'){
            return res.status(404).json({ msg: 'Profile not found'});
        }
        res.status(500).send('Server Error');
    }
});

//@route    DELETE api/profile
//@desc     Delete profile, user & posts
//@access   Private
router.delete('/', auth,async (req, res)=>{
    try {
        // Remove User's post
        await Post.deleteMany({ user: req.user.id });
        // Remove profile
        await Profile.findOneAndRemove({user: req.user.id});
        // Remove user
        await User.findOneAndRemove({_id: req.user.id});
        res.json({msg: 'User Deleted'});
    }catch (e) {
        console.error(e.message);
        res.status(500).send('Server Error');
    }
});

//@route    PUT api/profile/experience
//@desc     Add profile experience
//@access   Private
router.put('/experience',
    [   auth,
        check('title', 'Title is required').not().isEmpty(),
        check('company', 'Company is required').not().isEmpty(),
        check('from', 'From Date is required').not().isEmpty()
    ], async (req, res)=>{

        // Verify Middleware Errors
        const errors = validationResult(req);
        if(!errors.isEmpty()){
            return res.status(400).json({errors: errors.array()});
        }

        // Create a new Object for experience
        const { title, company, location, from, to, current, description } = req.body;

        const newExp = { title, company, location, from, to, current, description };

        try{
            const profile = await Profile.findOne({ user: req.user.id });
            profile.experience.unshift(newExp);
            await profile.save();

            res.json(profile);
        }catch (e) {
            console.log(e.message);
            res.status(500).send('Server Error');
        }

});

//@route    DELETE api/profile/experience/:exp_id
//@desc     Delete experience from profile
//@access   Private
router.delete('/experience/:exp_id', auth, async (req, res) => {
   try{
       // Find Profile and remove experience by _id
       const profile = await Profile.findOneAndUpdate(
           { user: req.user.id },
           {$pull: {experience: { _id: req.params.exp_id }}},
           {new: true}
           );

       return res.json(profile);

   } catch (e) {
       console.log(e.message);
       res.status(500).send('Server Error');
   }
});


//@route    PUT api/profile/education
//@desc     Add profile education
//@access   Private
router.put('/education',
    [   auth,
        check('school', 'School is required').not().isEmpty(),
        check('degree', 'Degree is required').not().isEmpty(),
        check('from', 'From Date is required').not().isEmpty(),
        check('fieldofstudy', 'Field of study is required').not().isEmpty()
    ], async (req, res)=>{

        // Verify Middleware Errors
        const errors = validationResult(req);
        if(!errors.isEmpty()){
            return res.status(400).json({errors: errors.array()});
        }

        // Create a new Object for experience
        const { school, degree, fieldofstudy, from, to, current, description } = req.body;

        const newEdu = {  school, degree, fieldofstudy, from, to, current, description };

        try{
            const profile = await Profile.findOne({ user: req.user.id });
            profile.education.unshift(newEdu);
            await profile.save();

            res.json(profile);
        }catch (e) {
            console.log(e.message);
            res.status(500).send('Server Error');
        }

    });

//@route    DELETE api/profile/education/:edu_id
//@desc     Delete education from profile
//@access   Private
router.delete('/education/:edu_id', auth, async (req, res) => {
    try{
        // Find Profile and remove experience by _id
        const profile = await Profile.findOneAndUpdate(
            { user: req.user.id },
            {$pull: {education: { _id: req.params.edu_id }}},
            {new: true}
        );

        return res.json(profile);

    } catch (e) {
        console.log(e.message);
        res.status(500).send('Server Error');
    }
});

//@route    GET api/profile/github/:username
//@desc     Get user repos from Github
//@access   Public
router.get('/github/:username', (req, res)=>{
    try{
        const options = {
            uri: `https://api.github.com/users/${req.params.username}/repos?per_page=5&sort=created:asc&client_id=${config.get('githubClentId')}&client_secret=${config.get('githubSecret')}`,
            method: 'GET',
            headers: {'user-agent': 'node.js'}
        };


        request(options, (error, response, body)=>{
            if(error){
                console.error(error);
            }
            if(response.statusCode !== 200){
                return res.status(404).json({msg: 'No Github profile found'})
            }
            res.json(JSON.parse(body));
        });

    }catch (e) {
        console.error(e.message);
        res.status(500).send('Server Error');
    }
});

module.exports = router;